Fix error while trying to delete not connected user

+++ listenbrainz/webserver/views/profile.py
@@ -347,7 +347,12 @@ def refresh_service_token(service_name: str):
 @api_login_required
 def music_services_disconnect(service_name: str):
     service = _get_service_or_raise_404(service_name)
-    service.remove_user(current_user.id)
+    user = service.get_user(current_user.id)
+    # this is to support the workflow of changing permissions in a single step
+    # we delete the current permissions and then try to authenticate with new ones
+    # we should try to delete the current permissions only if the user has connected previously
+    if user:
+        service.remove_user(current_user.id)
 
     action = request.form.get(service_name)
     if not action or action == 'disable':
